Creation of MecanumKimematicModel class and the movement of kinematic functions to it from Util. Docs coming soon
