+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms/Drivetrain/Utils/Utils.java
@@ -2,9 +2,7 @@ package org.firstinspires.ftc.teamcode.Mechanisms.Drivetrain.Utils;
 
 import org.ejml.simple.SimpleMatrix;
 public class Utils {
-    public static double r = 2.16535; //in
-    public static double l = 5.7; //in
-    public static double w = 5.31496; //in
+
     public static SimpleMatrix R(double th) {
         SimpleMatrix rotation = new SimpleMatrix(
                 new double[][]{
@@ -15,14 +13,6 @@ public class Utils {
         );
         return rotation;
     }
-    static SimpleMatrix H_inv = new SimpleMatrix(
-            new double[][]{
-                    new double[]{1d, -1d, -(l + w)},
-                    new double[]{1d, 1d, -(l + w)},
-                    new double[]{1d, -1d, (l + w)},
-                    new double[]{1d, 1d, (l + w)}
-            }
-    ); //Inverse kinematics Matrix
     public static SimpleMatrix rotateBodyToGlobal(SimpleMatrix vectorBody, double th) {
         return R(th).mult(vectorBody);
     }
@@ -30,10 +20,7 @@ public class Utils {
     public static SimpleMatrix rotateGlobalToBody(SimpleMatrix vectorGlobal, double th) {
         return R(th).invert().mult(vectorGlobal);
     }
-    public static SimpleMatrix inverseKinematics(SimpleMatrix twist) {
-        SimpleMatrix wheelSpeeds = H_inv.scale(1 / r).mult(twist);
-        return wheelSpeeds;
-    }
+
     public static double angleWrap(double radians) {
 
         while (radians > Math.PI) {
